backend/requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.2.5
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
Werkzeug==2.2.2
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
apiflask @ git+https://github.com/apiflask/apiflask.git@main